# SurveyOnline.API/bin/Debug/net5.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Survey Online",
    "version": "v1"
  },
  "paths": {
...
    "/api/Answers/{answerId}": {
      "get": {
        "tags": [
          "Answers"
        ],
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Answers"
        ],
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Answers"
        ],
        "parameters": [
          {
            "name": "answerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Surveies/getDetail/{surveyId}/user/{userId}": {
      "get": {
        "tags": [
          "Surveies"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "surveyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/UserAnswers": {
      "post": {
        "tags": [
          "UserAnswers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAnswerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAnswerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users/Register": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users/authenticate": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users/ConfirmEmail": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnswerUpdateRequest": {
        "required": [
          "Content",
          "QuestionId"
        ],
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "QuestionId": {
            "type": "integer",
            "format": "int32"
          },
          "Content": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UserAnswerCreateRequest": {
        "required": [
          "QuestionId",
          "SurveyId",
          "UserId"
        ],
        "type": "object",
        "properties": {
          "AnswerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "SurveyId": {
            "type": "integer",
            "format": "int32"
          },
          "UserId": {
            "type": "string"
          },
          "QuestionId": {
            "type": "integer",
            "format": "int32"
          },
          "Response": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserAnswerDto": {
        "type": "object",
        "properties": {
          "SurveyId": {
            "type": "integer",
            "format": "int32"
          },
          "Answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAnswerCreateRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegisterUserRequest": {
        "required": [
          "ConfirmPassword",
          "Email",
          "FullName",
          "Password",
          "Role",
          "UserName"
        ],
        "type": "object",
        "properties": {
          "Email": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string",
            "format": "email"
          },
          "UserName": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "FullName": {
            "type": "string"
          },
          "Role": {
            "type": "integer",
            "format": "int32"
          },
          "Password": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "ConfirmPassword": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AuthenticateUserRequest": {
        "required": [
          "Password",
          "Username"
        ],
        "type": "object",
        "properties": {
          "Username": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "Password": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.\r\n                      Enter 'Bearer' [space] and then your token in the text input below.\r\n                      Example: 'Bearer 12345abcdef'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}